refactor ProjectsQueryFlags out of errors_replacer
